Add Solana support (#301)
* Pretending to be phantom
* Connect wallet and sign transaction
* Update isConnected logic
* signTransaction payload as an array
* add signAllTransactions
* injecting both providers
* move eth provider out of index and fix existed test
* Add solanaPubkey
* move web3 to dev dependency
* add base58 encoding for signing
* update signing
* Verify signatures
* Passing transaction to sign as base58 encoded string
* Move sendResponse to the base provider
* handle disconnect message
* requestAccounts call
* Add ProviderNetwork type, post network from js
* Fix Ethereum tests and add one solana test
* move isTrust / isDebug to base provider
* Add signMessage to solana provider, fix Scaffold connect issue
* Fix solana sign message
* sendAndConfirmRawTransaction
* Move sending logic to signAndSendTransaction only
* Refactor config
* Refactor ProviderNetwork
* Implemented signRawTransaction
* Fix tests
* Fix tx signing with solana Scaffold
* move sendError to base
* Send transactions from js instead of native
* Fix signAndSendTransaction
* Remove FIXME
* read isPhantom from config
