+++ README.md
@@ -5,7 +5,10 @@
 [![Platform](https://img.shields.io/cocoapods/p/TrustWeb3Provider.svg?style=flat)](http://cocoapods.org/pods/TrustWeb3Provider)
 [![Platform](https://img.shields.io/badge/platform-android-lightgrey.svg)](https://jitpack.io/#TrustWallet/trust-web3-provider/0.2.1)
 
-TrustWeb3Provider currently bundles [web3 0.20.x](https://github.com/trustwallet/trust-web3-provider/blob/master/src/package.json#L22), we will follow MetaMask proposal: [No Longer Injecting web3.js](https://medium.com/metamask/no-longer-injecting-web3-js-4a899ad6e59e).
+TrustWeb3Provider is multi-network web3 provider used by TrustWallet. Currently it supports:
+
+- Ethereum
+- Solana
 
 ## How to Identify Trust Provider
 
@@ -13,6 +16,8 @@ If trust provider injected properly `isTrust` will be `true`
 
 ```javascript
 window.ethereum.isTrust
+// or
+window.solana.isTrust
 ```
 
 ## Installation
@@ -69,6 +74,7 @@ dependencies {
 [vikmeup](https://github.com/vikmeup)  
 [hewigovens](https://github.com/hewigovens)  
 [madcake](https://github.com/madcake)  
+[rsrbk](https://github.com/rsrbk)
 
 ## License
