+++ android/app/src/main/java/com/trust/web3/demo/MainActivity.kt
@@ -56,7 +56,7 @@ class MainActivity : AppCompatActivity() {
                 isDebug: true
             };
             window.ethereum = new trustwallet.Provider(config);
-            window.web3 = new trustwallet.Web3(window.ethereum);
+            window.solana = new trustwallet.SolanaProvider(config);
             trustwallet.postMessage = (json) => {
                 window._tw_.postMessage(JSON.stringify(json));
             }
