+++ ios/TrustWeb3Provider/DAppMethod.swift
@@ -7,10 +7,10 @@
 import Foundation
 
 enum DAppMethod: String, Decodable, CaseIterable {
-    case signTransaction
-    case signPersonalMessage
+    case signRawTransaction
     case signMessage
     case signTypedMessage
+    case signPersonalMessage
     case ecRecover
     case requestAccounts
     case watchAsset
