+++ src/index.js
@@ -6,386 +6,11 @@
 
 "use strict";
 
-import Web3 from "web3";
-import RPCServer from "./rpc";
-import ProviderRpcError from "./error";
-import Utils from "./utils";
-import IdMapping from "./id_mapping";
-import { EventEmitter } from "events";
-import isUtf8 from "isutf8";
-import { TypedDataUtils, SignTypedDataVersion } from "@metamask/eth-sig-util";
-
-class TrustWeb3Provider extends EventEmitter {
-  constructor(config) {
-    super();
-    this.setConfig(config);
-
-    this.idMapping = new IdMapping();
-    this.callbacks = new Map();
-    this.wrapResults = new Map();
-    this.isTrust = true;
-    this.isMetaMask = false;
-    this.isDebug = !!config.isDebug;
-
-    this.emitConnect(this.chainId);
-  }
-
-  setAddress(address) {
-    const lowerAddress = (address || "").toLowerCase();
-    this.address = lowerAddress;
-    this.ready = !!address;
-    for (var i = 0; i < window.frames.length; i++) {
-      const frame = window.frames[i];
-      if (frame.ethereum && frame.ethereum.isTrust) {
-        frame.ethereum.address = lowerAddress;
-        frame.ethereum.ready = !!address;
-      }
-    }
-  }
-
-  setConfig(config) {
-    this.setAddress(config.address);
-
-    this.networkVersion = "" + config.chainId;
-    this.chainId = "0x" + (config.chainId || 1).toString(16);
-    this.rpc = new RPCServer(config.rpcUrl);
-    this.isDebug = !!config.isDebug;
-  }
-
-  request(payload) {
-    // this points to window in methods like web3.eth.getAccounts()
-    var that = this;
-    if (!(this instanceof TrustWeb3Provider)) {
-      that = window.ethereum;
-    }
-    return that._request(payload, false);
-  }
-
-  /**
-   * @deprecated Listen to "connect" event instead.
-   */
-  isConnected() {
-    return true;
-  }
-
-  /**
-   * @deprecated Use request({method: "eth_requestAccounts"}) instead.
-   */
-  enable() {
-    console.log(
-      'enable() is deprecated, please use window.ethereum.request({method: "eth_requestAccounts"}) instead.'
-    );
-    return this.request({ method: "eth_requestAccounts", params: [] });
-  }
-
-  /**
-   * @deprecated Use request() method instead.
-   */
-  send(payload) {
-    if (this.isDebug) {
-      console.log(`==> send payload ${JSON.stringify(payload)}`);
-    }
-    let response = { jsonrpc: "2.0", id: payload.id };
-    switch (payload.method) {
-      case "eth_accounts":
-        response.result = this.eth_accounts();
-        break;
-      case "eth_coinbase":
-        response.result = this.eth_coinbase();
-        break;
-      case "net_version":
-        response.result = this.net_version();
-        break;
-      case "eth_chainId":
-        response.result = this.eth_chainId();
-        break;
-      default:
-        throw new ProviderRpcError(
-          4200,
-          `Trust does not support calling ${payload.method} synchronously without a callback. Please provide a callback parameter to call ${payload.method} asynchronously.`
-        );
-    }
-    return response;
-  }
-
-  /**
-   * @deprecated Use request() method instead.
-   */
-  sendAsync(payload, callback) {
-    console.log(
-      "sendAsync(data, callback) is deprecated, please use window.ethereum.request(data) instead."
-    );
-    // this points to window in methods like web3.eth.getAccounts()
-    var that = this;
-    if (!(this instanceof TrustWeb3Provider)) {
-      that = window.ethereum;
-    }
-    if (Array.isArray(payload)) {
-      Promise.all(payload.map((_payload) => that._request(_payload)))
-        .then((data) => callback(null, data))
-        .catch((error) => callback(error, null));
-    } else {
-      that
-        ._request(payload)
-        .then((data) => callback(null, data))
-        .catch((error) => callback(error, null));
-    }
-  }
-
-  /**
-   * @private Internal rpc handler
-   */
-  _request(payload, wrapResult = true) {
-    this.idMapping.tryIntifyId(payload);
-    if (this.isDebug) {
-      console.log(`==> _request payload ${JSON.stringify(payload)}`);
-    }
-    this.fillJsonRpcVersion(payload);
-    return new Promise((resolve, reject) => {
-      if (!payload.id) {
-        payload.id = Utils.genId();
-      }
-      this.callbacks.set(payload.id, (error, data) => {
-        if (error) {
-          reject(error);
-        } else {
-          resolve(data);
-        }
-      });
-      this.wrapResults.set(payload.id, wrapResult);
-
-      switch (payload.method) {
-        case "eth_accounts":
-          return this.sendResponse(payload.id, this.eth_accounts());
-        case "eth_coinbase":
-          return this.sendResponse(payload.id, this.eth_coinbase());
-        case "net_version":
-          return this.sendResponse(payload.id, this.net_version());
-        case "eth_chainId":
-          return this.sendResponse(payload.id, this.eth_chainId());
-        case "eth_sign":
-          return this.eth_sign(payload);
-        case "personal_sign":
-          return this.personal_sign(payload);
-        case "personal_ecRecover":
-          return this.personal_ecRecover(payload);
-        case "eth_signTypedData_v3":
-          return this.eth_signTypedData(payload, SignTypedDataVersion.V3);
-        case "eth_signTypedData":
-        case "eth_signTypedData_v4":
-          return this.eth_signTypedData(payload, SignTypedDataVersion.V4);
-        case "eth_sendTransaction":
-          return this.eth_sendTransaction(payload);
-        case "eth_requestAccounts":
-          return this.eth_requestAccounts(payload);
-        case "wallet_watchAsset":
-          return this.wallet_watchAsset(payload);
-        case "wallet_addEthereumChain":
-          return this.wallet_addEthereumChain(payload);
-        case "wallet_switchEthereumChain":
-          return this.wallet_switchEthereumChain(payload);
-        case "eth_newFilter":
-        case "eth_newBlockFilter":
-        case "eth_newPendingTransactionFilter":
-        case "eth_uninstallFilter":
-        case "eth_subscribe":
-          throw new ProviderRpcError(
-            4200,
-            `Trust does not support calling ${payload.method}. Please use your own solution`
-          );
-        default:
-          // call upstream rpc
-          this.callbacks.delete(payload.id);
-          this.wrapResults.delete(payload.id);
-          return this.rpc
-            .call(payload)
-            .then((response) => {
-              if (this.isDebug) {
-                console.log(`<== rpc response ${JSON.stringify(response)}`);
-              }
-              wrapResult ? resolve(response) : resolve(response.result);
-            })
-            .catch(reject);
-      }
-    });
-  }
-
-  fillJsonRpcVersion(payload) {
-    if (payload.jsonrpc === undefined) {
-      payload.jsonrpc = "2.0";
-    }
-  }
-
-  emitConnect(chainId) {
-    this.emit("connect", { chainId: chainId });
-  }
-
-  emitChainChanged(chainId) {
-    this.emit("chainChanged", chainId);
-    this.emit("networkChanged", chainId);
-  }
-
-  eth_accounts() {
-    return this.address ? [this.address] : [];
-  }
-
-  eth_coinbase() {
-    return this.address;
-  }
-
-  net_version() {
-    return this.networkVersion;
-  }
-
-  eth_chainId() {
-    return this.chainId;
-  }
-
-  eth_sign(payload) {
-    const buffer = Utils.messageToBuffer(payload.params[1]);
-    const hex = Utils.bufferToHex(buffer);
-    if (isUtf8(buffer)) {
-      this.postMessage("signPersonalMessage", payload.id, { data: hex });
-    } else {
-      this.postMessage("signMessage", payload.id, { data: hex });
-    }
-  }
-
-  personal_sign(payload) {
-    const message = payload.params[0];
-    const buffer = Utils.messageToBuffer(message);
-    if (buffer.length === 0) {
-      // hex it
-      const hex = Utils.bufferToHex(message);
-      this.postMessage("signPersonalMessage", payload.id, { data: hex });
-    } else {
-      this.postMessage("signPersonalMessage", payload.id, { data: message });
-    }
-  }
-
-  personal_ecRecover(payload) {
-    this.postMessage("ecRecover", payload.id, {
-      signature: payload.params[1],
-      message: payload.params[0],
-    });
-  }
-
-  eth_signTypedData(payload, version) {
-    const message = JSON.parse(payload.params[1]);
-    const hash = TypedDataUtils.eip712Hash(message, version);
-    this.postMessage("signTypedMessage", payload.id, {
-      data: "0x" + hash.toString("hex"),
-      raw: payload.params[1],
-    });
-  }
-
-  eth_sendTransaction(payload) {
-    this.postMessage("signTransaction", payload.id, payload.params[0]);
-  }
-
-  eth_requestAccounts(payload) {
-    this.postMessage("requestAccounts", payload.id, {});
-  }
-
-  wallet_watchAsset(payload) {
-    let options = payload.params.options;
-    this.postMessage("watchAsset", payload.id, {
-      type: payload.type,
-      contract: options.address,
-      symbol: options.symbol,
-      decimals: options.decimals || 0,
-    });
-  }
-
-  wallet_addEthereumChain(payload) {
-    this.postMessage("addEthereumChain", payload.id, payload.params[0]);
-  }
-
-  wallet_switchEthereumChain(payload) {
-    this.postMessage("switchEthereumChain", payload.id, payload.params[0]);
-  }
-
-  /**
-   * @private Internal js -> native message handler
-   */
-  postMessage(handler, id, data) {
-    if (this.ready || handler === "requestAccounts") {
-      let object = {
-        id: id,
-        name: handler,
-        object: data,
-      };
-      if (window.trustwallet.postMessage) {
-        window.trustwallet.postMessage(object);
-      } else {
-        // old clients
-        window.webkit.messageHandlers[handler].postMessage(object);
-      }
-    } else {
-      // don't forget to verify in the app
-      this.sendError(id, new ProviderRpcError(4100, "provider is not ready"));
-    }
-  }
-
-  /**
-   * @private Internal native result -> js
-   */
-  sendResponse(id, result) {
-    let originId = this.idMapping.tryPopId(id) || id;
-    let callback = this.callbacks.get(id);
-    let wrapResult = this.wrapResults.get(id);
-    let data = { jsonrpc: "2.0", id: originId };
-    if (
-      result !== null &&
-      typeof result === "object" &&
-      result.jsonrpc &&
-      result.result
-    ) {
-      data.result = result.result;
-    } else {
-      data.result = result;
-    }
-    if (this.isDebug) {
-      console.log(
-        `<== sendResponse id: ${id}, result: ${JSON.stringify(
-          result
-        )}, data: ${JSON.stringify(data)}`
-      );
-    }
-    if (callback) {
-      wrapResult ? callback(null, data) : callback(null, result);
-      this.callbacks.delete(id);
-    } else {
-      console.log(`callback id: ${id} not found`);
-      // check if it's iframe callback
-      for (var i = 0; i < window.frames.length; i++) {
-        const frame = window.frames[i];
-        try {
-          if (frame.ethereum.callbacks.has(id)) {
-            frame.ethereum.sendResponse(id, result);
-          }
-        } catch (error) {
-          console.log(`send response to frame error: ${error}`);
-        }
-      }
-    }
-  }
-
-  /**
-   * @private Internal native error -> js
-   */
-  sendError(id, error) {
-    console.log(`<== ${id} sendError ${error}`);
-    let callback = this.callbacks.get(id);
-    if (callback) {
-      callback(error instanceof Error ? error : new Error(error), null);
-      this.callbacks.delete(id);
-    }
-  }
-}
+import TrustWeb3Provider from "./ethereum_provider";
+import TrustSolanaWeb3Provider from "./solana_provider";
 
 window.trustwallet = {
   Provider: TrustWeb3Provider,
-  Web3: Web3,
+  SolanaProvider: TrustSolanaWeb3Provider,
   postMessage: null,
 };
