+++ src/package.json
@@ -12,17 +12,19 @@
   "scripts": {
     "build": "npx browserify index.js -t [ babelify --presets [ @babel/preset-env ] ] | uglifyjs > ../dist/trust-min.js",
     "lint": "eslint . --fix",
-    "test": "jest"
+    "test": "jest",
+    "lfs-fix": "git rm --cached --force ../dist/trust-min.js && git add ../dist/trust-min.js"
   },
   "engines": {
     "node": ">=14"
   },
   "dependencies": {
     "@metamask/eth-sig-util": "^4.0.1",
+    "@solana/web3.js": "^1.5.0",
+    "bs58": "^4.0.1",
     "buffer": "^5.6.0",
     "events": "^3.2.0",
-    "isutf8": "^3.1.1",
-    "web3": "^0.20.7"
+    "isutf8": "^3.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.16.12",
@@ -33,6 +35,7 @@
     "ethereumjs-util": "^7.0.5",
     "jest": "^27.4.7",
     "uglify-js": "^3.15.0",
+    "web3": "^0.20.7",
     "whatwg-fetch": "^3.6.2"
   },
   "jest": {
