+++ src/utils.js
@@ -43,7 +43,7 @@ class Utils {
   static messageToBuffer(message) {
     var buffer = Buffer.from([]);
     try {
-      if ((typeof (message) === "string")) {
+      if (typeof message === "string") {
         buffer = Buffer.from(message.replace("0x", ""), "hex");
       } else {
         buffer = Buffer.from(message);
